+++ xmldb_lib/VectorInt.h
@@ -4,19 +4,19 @@
 #include<memory.h>
 
 struct VectorInt{
-    int length;
-    int size;
+    size_t length;
+    size_t size;
     int* items;
 };
 
-void init_VectorInt(struct VectorInt* v,int size);
-void VectorInt_Resize(struct VectorInt *vect,int new_size);
+void init_VectorInt(struct VectorInt* v,size_t size);
+void VectorInt_Resize(struct VectorInt *vect,size_t new_size);
 void free_VectorInt(struct VectorInt *vect);
 void clear_VectorInt(struct VectorInt *vect);
 void appendto_VectorInt(struct VectorInt *src_dest, int value);
 void concatenate_VectorInt(struct VectorInt *dest, struct VectorInt* src);
-void inserto_VectorInt(struct VectorInt *src_dest, int index, int value);
-void removefrom_VectorInt(struct VectorInt *src_dest, int index);
-void Slice_VectorInt(struct VectorInt *Result, struct VectorInt *input, int from, int to);
+void inserto_VectorInt(struct VectorInt *src_dest, size_t index, int value);
+void removefrom_VectorInt(struct VectorInt *src_dest, size_t index);
+void Slice_VectorInt(struct VectorInt *Result, struct VectorInt *input, size_t from, size_t to);
 void free_VectorIntReturn(struct VectorInt *vect);
 #endif
